Added ---HumanList Folder to Track by Titles Updated channelwatch/AutoDL.sh - now Automaticly updates ---HumanList   and writes the New Video Titles into ___new.txt - to reset mannually remove AllKnownVids.list

+++ _deleteNoVid.sh
@@ -1,4 +1,6 @@
 #!/bin/sh
+#  	Parameters:
+# 1	Keep [Optiona] "keep" to keep the list
 
 while read -r line
 do
@@ -6,4 +8,4 @@ do
   rm -r -f -- $(echo "$line" | sed 's/\///')
 done < noVid.list
 
-rm -f noVid.list
+[ "$1" = "keep" ] || rm -f noVid.list
